Format broad kernel subsystems under a truncated diff
The T1 diff is truncated, so this is described as a formatting pass across CPU, disk, install, NVMe, ACPI, and related kernel files.

The T1 diff for this commit is truncated, so this message keeps the claim at the file/stat level instead of pretending the internal rewrite is fully visible. This is integration pressure around a networked interactive runtime: AI HTTPS, timers, DNS/adapters, HID, Pixi, boot benchmarks, and faster display paths are made to coexist.

+++ src/usb/input.rs
@@ -103,11 +103,11 @@ static QJS_MOUSE_PIPE: Mutex<QjsMousePipe> = Mutex::new(QjsMousePipe::new());
 #[inline]
 fn now_ms() -> u64 {
     let ticks = now();
-    let hz = TICK_HZ as u64;
+    let hz = TICK_HZ;
     if hz == 0 {
         return 0;
     }
-    (ticks as u64).saturating_mul(1000) / hz
+    ticks.saturating_mul(1000) / hz
 }
 
 #[derive(Copy, Clone, Debug)]
